+++ src/SqlArtisan/Internal/SqlBuilder/Update/UpdateBuilder.cs
@@ -1,7 +1,12 @@
 namespace SqlArtisan.Internal;
 
-internal sealed class UpdateBuilder(DbTableBase table, params SqlPart[] rootParts) :
+internal sealed class UpdateBuilder(DbTableBase table, DmlJoinState state, params SqlPart[] rootParts) :
     SqlBuilderBase(rootParts),
+    IUpdateBuilderFrom,
+    IUpdateBuilderFromJoinOn,
+    IUpdateBuilderJoinOn,
+    IUpdateBuilderJoined,
+    IUpdateBuilderJoinedSet,
     IUpdateBuilderSet,
     IUpdateBuilderUpdate,
     IUpdateBuilderWhere
@@ -17,12 +22,107 @@ internal sealed class UpdateBuilder(DbTableBase table, params SqlPart[] rootPart
     public SqlStatement Build(Dbms dbms) =>
         BuildCore(dbms);
 
+    public IUpdateBuilderFrom From(params TableReference[] tables)
+    {
+        AddPart(new FromClause(tables));
+        state.HasFrom = true;
+
+        foreach (TableReference reference in tables)
+        {
+            if (ReferenceEquals(reference, table))
+            {
+                state.TargetRepeatedInFrom = true;
+                break;
+            }
+        }
+
+        return this;
+    }
+
+    public IUpdateBuilderJoinOn FullJoin(TableReference joined)
+    {
+        AddJoin(new FullJoinClause(joined));
+        return this;
+    }
+
+    IUpdateBuilderFromJoinOn IUpdateBuilderFrom.FullJoin(TableReference joined)
+    {
+        AddJoin(new FullJoinClause(joined));
+        return this;
+    }
+
+    public IUpdateBuilderJoinOn InnerJoin(TableReference joined)
+    {
+        AddJoin(new InnerJoinClause(joined));
+        return this;
+    }
+
+    IUpdateBuilderFromJoinOn IUpdateBuilderFrom.InnerJoin(TableReference joined)
+    {
+        AddJoin(new InnerJoinClause(joined));
+        return this;
+    }
+
+    public IUpdateBuilderJoinOn LeftJoin(TableReference joined)
+    {
+        AddJoin(new LeftJoinClause(joined));
+        return this;
+    }
+
+    IUpdateBuilderFromJoinOn IUpdateBuilderFrom.LeftJoin(TableReference joined)
+    {
+        AddJoin(new LeftJoinClause(joined));
+        return this;
+    }
+
+    IUpdateBuilderFrom IUpdateBuilderFromJoinOn.On(SqlCondition condition)
+    {
+        AddPart(new OnClause(condition));
+        return this;
+    }
+
+    IUpdateBuilderJoined IUpdateBuilderJoinOn.On(SqlCondition condition)
+    {
+        AddPart(new OnClause(condition));
+        return this;
+    }
+
     public IReturningBuilder Returning(params object[] expressions) =>
         ReturningBuilder.Create(this, expressions);
 
+    public IUpdateBuilderJoinOn RightJoin(TableReference joined)
+    {
+        AddJoin(new RightJoinClause(joined));
+        return this;
+    }
+
+    IUpdateBuilderFromJoinOn IUpdateBuilderFrom.RightJoin(TableReference joined)
+    {
+        AddJoin(new RightJoinClause(joined));
+        return this;
+    }
+
     public IUpdateBuilderSet Set(params EqualityBasedCondition[] assignments)
     {
-        AddPart(UpdateSetClause.Parse(assignments));
+        AddPart(UpdateSetClause.Parse(assignments, state));
+        return this;
+    }
+
+    IUpdateBuilderJoinedSet IUpdateBuilderJoined.Set(params EqualityBasedCondition[] assignments)
+    {
+        AddPart(UpdateSetClause.Parse(assignments, state));
+        return this;
+    }
+
+    IUpdateBuilderFrom IUpdateBuilderFromJoinOn.Using(DbColumn column, params DbColumn[] additionalColumns)
+    {
+        AddPart(new JoinUsingClause([column, .. additionalColumns]));
+        return this;
+    }
+
+    IUpdateBuilderJoined IUpdateBuilderJoinOn.Using(DbColumn column, params DbColumn[] additionalColumns)
+    {
+        AddPart(new JoinUsingClause([column, .. additionalColumns]));
         return this;
     }
 
@@ -32,6 +132,22 @@ internal sealed class UpdateBuilder(DbTableBase table, params SqlPart[] rootPart
         return this;
     }
 
-    protected override void Validate(Dbms dbms) =>
-        DmlTargetGuard.ThrowIfAliasedOnSqlServer(table, dbms);
+    protected override void Validate(Dbms dbms)
+    {
+        if (state.IsJoined)
+        {
+            DmlTargetGuard.ThrowIfJoinedTargetUnaliased(table);
+            DmlTargetGuard.ThrowIfJoinedUpdateShapeInvalid(state, dbms);
+        }
+        else
+        {
+            DmlTargetGuard.ThrowIfAliasedOnSqlServer(table, dbms);
+        }
+    }
+
+    private void AddJoin(SqlPart joinClause)
+    {
+        AddPart(joinClause);
+        state.HasJoin = true;
+    }
 }